@@ -9,9 +9,9 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
+    "@types/node": "^18.16.18",
+    "@types/react": "^18.2.12",
+    "@types/react-dom": "^18.2.5",
     "@vitejs/plugin-react": "^4.2.1",
     "animate.css": "^4.1.1",
     "axios": "^1.1.3",
@@ -22,7 +22,7 @@
     "js-cookie": "^3.0.1",
     "pluralize": "^8.0.0",
     "react": "^18.2.0",
-    "react-animate-height": "^3.1.0",
+    "react-animate-height": "^3.1.1",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-file-base64": "^1.0.3",
@@ -31,10 +31,10 @@
     "react-router-dom": "^6.4.2",
     "react-toastify": "^9.0.8",
     "recharts": "^2.1.16",
-    "swiper": "^8.4.5",
+    "swiper": "^8.4.7",
     "vite": "^5.1.5",
     "vite-tsconfig-paths": "^4.3.1",
-    "web-vitals": "^3.1.1",
+    "web-vitals": "^3.3.2",
     "workbox-background-sync": "^6.5.4",
     "workbox-broadcast-update": "^6.5.4",
     "workbox-cacheable-response": "^6.5.4",
@@ -75,25 +75,26 @@
       "last 1 safari version"
     ]
   },
+  "type": "module",
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/js-cookie": "^3.0.2",
     "@types/pluralize": "^0.0.29",
-    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.11",
     "autoprefixer": "^10.4.12",
     "concurrently": "^7.6.0",
-    "eslint": "^8.0.1",
+    "eslint": "^8.25.0",
     "eslint-config-standard-with-typescript": "^24.0.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.5.1",
+    "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-simple-import-sort": "^8.0.0",
     "openapi-typescript-codegen": "^0.23.0",
-    "postcss": "^8.4.21",
+    "postcss": "^8.4.24",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.2.4",
-    "typescript": "*"
+    "tailwindcss": "^3.3.2",
+    "typescript": "^4.8.4"
   }
 }
