@@ -2,63 +2,66 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { ToastContainer } from 'react-toastify';
 
-const Home = lazy(() => import('./pages/Home'));
-const Wishlist = lazy(() => import('./pages/Wishlist'));
-const Checkout = lazy(() => import('./pages/Checkout'));
-const Custom404 = lazy(() => import('./pages/404'));
-const ComingSoon = lazy(() => import('./pages/ComingSoon'));
-const Products = lazy(() => import('./pages/Products'));
-const Product = lazy(() => import('./pages/Product'));
-const Profile = lazy(() => import('./pages/Profile'));
-const Admin = lazy(() => import('./pages/Admin'));
+// const Home = lazy(() => import('./pages/Home'));
+// const Wishlist = lazy(() => import('./pages/Wishlist'));
+// const Checkout = lazy(() => import('./pages/Checkout'));
+// const Custom404 = lazy(() => import('./pages/404'));
+// const ComingSoon = lazy(() => import('./pages/ComingSoon'));
+// const Products = lazy(() => import('./pages/Products'));
+// const Product = lazy(() => import('./pages/Product'));
+// const Profile = lazy(() => import('./pages/Profile'));
+// const Admin = lazy(() => import('./pages/Admin'));
 
-// import Home from './pages/Home';
-// import Wishlist from './pages/Wishlist';
-// import Checkout from './pages/Checkout';
-// import Custom404 from './pages/404';
-// import ComingSoon from './pages/ComingSoon';
-// import Products from './pages/Products';
-// import Product from './pages/Product';
-// import Profile from './pages/Profile';
-// import Admin from './pages/Admin';
+import Home from './pages/Home';
+import Wishlist from './pages/Wishlist';
+import Checkout from './pages/Checkout';
+import Custom404 from './pages/404';
+import ComingSoon from './pages/ComingSoon';
+import Products from './pages/Products';
+import Product from './pages/Product';
+import Profile from './pages/Profile';
+import Admin from './pages/Admin';
 
 import Header from './components/Header';
 import Footer from './components/Footer';
 
-const Orders = lazy(() => import('./components/admin/Orders'));
-const Customers = lazy(() => import('./components/admin/Customers'));
-const Categories = lazy(() => import('./components/admin/Categories'));
-const Banners = lazy(() => import('./components/admin/Banners'));
-const Banner = lazy(() => import('./components/admin/Banner'));
-const CreateBanner = lazy(() => import('./components/admin/CreateBanner'));
-const AdminProducts = lazy(() => import('./components/admin/Products'));
-const AdminOrder = lazy(() => import('./components/admin/Order'));
-const AdminProduct = lazy(() => import('./components/admin/Product'));
-const Dashboard = lazy(() => import('./components/admin/Dashboard'));
-const PersonalData = lazy(() => import('./components/profile/PersonalData'));
-const Order = lazy(() => import('./components/profile/Order'));
-const Category = lazy(() => import('./components/admin/Category'));
-const Customer = lazy(() => import('./components/admin/Customer'));
-const Search = lazy(() => import('./components/Search'));
-const CreateCategory = lazy(() => import('./components/admin/CreateCategory'));
-const CreateProduct = lazy(() => import('./components/admin/CreateProduct'));
-const TestModel = lazy(() => import('./pages/TestModel'));
+// const Orders = lazy(() => import('./components/admin/Orders'));
+// const Customers = lazy(() => import('./components/admin/Customers'));
+// const Categories = lazy(() => import('./components/admin/Categories'));
+// const Banners = lazy(() => import('./components/admin/Banners'));
+// const Banner = lazy(() => import('./components/admin/Banner'));
+// const CreateBanner = lazy(() => import('./components/admin/CreateBanner'));
+// const AdminProducts = lazy(() => import('./components/admin/Products'));
+// const AdminOrder = lazy(() => import('./components/admin/Order'));
+// const AdminProduct = lazy(() => import('./components/admin/Product'));
+// const Dashboard = lazy(() => import('./components/admin/Dashboard'));
+// const PersonalData = lazy(() => import('./components/profile/PersonalData'));
+// const Order = lazy(() => import('./components/profile/Order'));
+// const Category = lazy(() => import('./components/admin/Category'));
+// const Customer = lazy(() => import('./components/admin/Customer'));
+// const Search = lazy(() => import('./components/Search'));
+// const CreateCategory = lazy(() => import('./components/admin/CreateCategory'));
+// const CreateProduct = lazy(() => import('./components/admin/CreateProduct'));
+// const TestModel = lazy(() => import('./pages/TestModel'));
 
-// import Orders from './components/admin/Orders';
-// import Customers from './components/admin/Customers';
-// import Categories from './components/admin/Categories';
-// import AdminProducts from './components/admin/Products';
-// import AdminOrder from './components/admin/Order';
-// import AdminProduct from './components/admin/Product';
-// import Dashboard from './components/admin/Dashboard';
-// import PersonalData from './components/profile/PersonalData';
-// import Order from './components/profile/Order';
-// import Category from './components/admin/Category';
-// import Customer from './components/admin/Customer';
-// import Search from './components/Search';
-// import CreateCategory from './components/admin/CreateCategory';
-// import CreateProduct from './components/admin/CreateProduct';
-// import TestModel from './pages/TestModel';
+import Banners from './components/admin/Banners';
+import Banner from './components/admin/Banner';
+import CreateBanner from './components/admin/CreateBanner';
+import Orders from './components/admin/Orders';
+import Customers from './components/admin/Customers';
+import Categories from './components/admin/Categories';
+import AdminProducts from './components/admin/Products';
+import AdminOrder from './components/admin/Order';
+import AdminProduct from './components/admin/Product';
+import Dashboard from './components/admin/Dashboard';
+import PersonalData from './components/profile/PersonalData';
+import Order from './components/profile/Order';
+import Category from './components/admin/Category';
+import Customer from './components/admin/Customer';
+import Search from './components/Search';
+import CreateCategory from './components/admin/CreateCategory';
+import CreateProduct from './components/admin/CreateProduct';
+import TestModel from './pages/TestModel';
 
 import UserRoutes from './util/UserRoutes';
 import AdminRoutes from './util/AdminRoutes';
